# types: improve StrToDate performance