### What problem does this PR solve?

Problem Summary: `StrToDate` is significantly slow due to the usage of regular expressions and maps.

### What is changed and how it works?

What's Changed:
- In `types/time.go`:
  - Parsing functions now use a new function, `digitPrefixLen`, to efficiently determine the number of digits in the next time component. This avoids using regular expressions.
  - The `ctx` object passed around function is now a custom structure `dateTimeParseContext` holding the smallest piece of information that is needed for later processing. This is faster to allocate and modify than `map[string]int`. Member types might be confusing, but hopefully it won't be a problem given they are documented and used in a relatively small scope.
- In `types/format_test.go`:
  - Added `BenchmarkStrToDate`.

How it Works:

```
name         old time/op  new time/op  delta
StrToDate-8  1.93µs ± 0%  0.78µs ± 0%   ~     (p=1.000 n=1+1)
```

### Related changes

- ~This PR is applied on top of https://github.com/pingcap/tidb/pull/17395 which should be merged before. Ping me if you prefer this PR to be rebased directly from master.~

### Check List

Tests

- Unit test

### Release note

* 2x performance improvement for `STR_TO_DATE`